src/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.3.0
cryptography==44.0.2
execnet==2.1.1
fastapi==0.117.1
future==1.0.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
iso8601==2.1.0
oauthlib==3.3.1
orjson==3.10.16
packaging==24.2
paramiko==3.5.1
pluggy==1.5.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
PyNaCl==1.5.0
pyparsing==3.2.3
pyserial==3.5
pytest==8.3.5
pytest-random-order==1.1.1
pytest-randomly==3.16.0
pytest-xdist==3.6.1
python-telegram-bot==22.4
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
serial==0.0.97
sniffio==1.3.1
starlette==0.48.0
termcolor==2.5.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.2.3
uvicorn==0.37.0
